Fix build export installation step
The recent refactoring of the documentation (see #1209) moved the
installation instructions for every individual build tool supported, so
that the directory structure was different than the one for package
managers.

Unfortunately, this broke the JS which displays the installation step
(the documentation for every build tool was no longer found).

This commit restores the previous directory structure, which is
unfortunately not as flat.

Fixes #1220